Conditional branches, instructions and jump tables
The loop statements, do-while, for and while compile to a series of
conditional jumps often reoreordered for optimization, this causes the
for loop to be a sugar coating for the while loop. In assembly, you can
even create new loop types and play till dawn which is really
interesting but they will mimic the general ones stated above hence
there inclusion in high level languages. The switch statement caught my
attention as it revealed that a hashtable is a jump table where the
index is generated by the hash function modulo size. The jump table is
way faster than chained if-else statements since the naive index
operation is usually addition or subtration.