Procedures and data transfer (variables)
More on registers, had a great time implementing a greeter application
in assembly, works like a charm. Now that my neovim config is competent
enough to to format and lsp the f*ck out of my projects, I can finally
use clangd and clang-format. As for the work, it was enjoyable, played a
little with the stack, callee- and caller-saved registers and the two:
rsp and rip.